pyproject.toml
[project]
name = "azure_calculator_backend"
version = "0.0.1"
description = "Azure Calculator backend API"
authors = [
    {name = "Shuo Lv", email = "claus.lv@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.27.0,<0.28.0",
    "pydantic[email]>=2.5.0,<2.6.0",
    "pydantic-settings>=2.1.0,<2.2.0",
    "sqlalchemy>=2.0.25,<2.1.0",
    "alembic>=1.13.1,<1.14.0",
    "asyncpg>=0.29.0,<0.30.0",
    "pymssql>=2.3.2,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<3.4.0",
    "passlib[bcrypt]>=1.7.4,<1.8.0",
    "python-multipart>=0.0.7,<0.1.0",
    "celery>=5.3.6,<5.4.0",
    "redis>=5.0.1,<5.1.0",
    "httpx>=0.26.0,<0.27.0",
    "tenacity>=8.2.3,<8.3.0",
    "loguru>=0.7.2,<0.8.0",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "openai (>=1.66.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<8.1.0",
    "pytest-asyncio>=0.23.3,<0.24.0",
    "pytest-cov>=4.1.0,<4.2.0",
    "black>=24.1.1,<24.2.0",
    "isort>=5.13.2,<5.14.0",
    "mypy>=1.8.0,<1.9.0",
    "ruff>=0.2.1,<0.3.0",
    "pre-commit>=3.6.0,<3.7.0",
]

[tool.poetry]
# 指定包目录
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "B", "C4", "SIM", "TID", "RUF"]
ignore = []
exclude = [
    ".git",
    ".ruff_cache",
    "__pycache__",
    "alembic",
    ".venv",
]